+++ .cspell.json
@@ -48,6 +48,7 @@
     "signups",
     "stylis",
     "tabler",
+    "tailwindcss",
     "tanstack",
     "timesheets",
     "tiptap",
