+++ packages/web/.eslintrc.json
@@ -27,6 +27,7 @@
     "coverage/",
     "next-env.d.ts",
     "next.config.js",
+    "tailwind.config.js",
     "postcss.config.js"
   ]
 }
