+++ packages/web/postcss.config.js
@@ -1,0 +1,10 @@
+const path = require("path");
+
+module.exports = {
+  plugins: {
+    tailwindcss: {
+      config: path.join(__dirname, "tailwind.config.js")
+    },
+    autoprefixer: {}
+  }
+};
